# specs/openapi/adscenter.yaml
openapi: 3.0.3
info:
  title: Adscenter Service
  version: 0.1.0
  license:
    name: Elastic-Internal
servers:
  - url: /
paths:
...
  /api/v1/adscenter/preflight:
    post:
      operationId: runPreflight
      summary: Run pre-flight diagnostics
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accountId:
                  type: string
                  description: Google Ads customer id (10 digits)
                validateOnly:
                  type: boolean
                  description: If true, skip live checks and only validate environment/config
                landingUrl:
                  type: string
                  description: Optional landing page URL to check reachability
              required: []
      responses:
        '200':
          description: Diagnostics result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreflightResult'
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PreflightCheck:
      type: object
      properties:
        code: { type: string }
        severity: { type: string, enum: [info, warn, error, skip] }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
      required: [code, severity, message]
    PreflightResult:
      type: object
      properties:
        summary: { type: string, enum: [ok, warn, error] }
        checks:
          type: array
          items: { $ref: '#/components/schemas/PreflightCheck' }
      required: [summary, checks]
...